[luci/export] Relocate optimize method with anonymous namespace
This will relocate optimize method with anonymous namespace into caller file.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
