--- compiler/luci/export/src/Optimize.h
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2020 Samsung Electronics Co., Ltd. All Rights Reserved
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef __OPTIMIZE_H__
-#define __OPTIMIZE_H__
-
-#include <loco.h>
-
-namespace luci
-{
-
-/**
- * @brief Run passes of graph transformations
- *
- */
-void optimize(loco::Graph *);
-
-} // namespace luci
-
-#endif // __OPTIMIZE_H__
